# PERF: slicing