--- a/pandas/core/indexes/base.py
+++ b/pandas/core/indexes/base.py
@@ -100,7 +100,6 @@ from pandas.core.dtypes.common import (
     is_float_dtype,
     is_hashable,
     is_integer,
-    is_integer_dtype,
     is_iterator,
     is_list_like,
     is_numeric_dtype,
@@ -162,7 +161,10 @@ from pandas.core.construction import (
     extract_array,
     sanitize_array,
 )
-from pandas.core.indexers import disallow_ndim_indexing
+from pandas.core.indexers import (
+    disallow_ndim_indexing,
+    is_valid_positional_slice,
+)
 from pandas.core.indexes.frozen import FrozenList
 from pandas.core.missing import clean_reindex_fill_method
 from pandas.core.ops import get_op_result_name
@@ -4061,7 +4063,7 @@ class Index(IndexOpsMixin, PandasObject):
         self._validate_indexer("positional", key.stop, "iloc")
         self._validate_indexer("positional", key.step, "iloc")
 
-    def _convert_slice_indexer(self, key: slice, kind: str_t):
+    def _convert_slice_indexer(self, key: slice, kind: Literal["loc", "getitem"]):
         """
         Convert a slice indexer.
 
@@ -4073,7 +4075,6 @@ class Index(IndexOpsMixin, PandasObject):
         key : label of the slice bound
         kind : {'loc', 'getitem'}
         """
-        assert kind in ["loc", "getitem"], kind
 
         # potentially cast the bounds to integers
         start, stop, step = key.start, key.stop, key.step
@@ -4086,10 +4087,21 @@ class Index(IndexOpsMixin, PandasObject):
             return self.slice_indexer(start, stop, step)
 
         # figure out if this is a positional indexer
-        def is_int(v):
-            return v is None or is_integer(v)
+        is_index_slice = is_valid_positional_slice(key)
 
-        is_index_slice = is_int(start) and is_int(stop) and is_int(step)
+        if kind == "getitem":
+            # called from the getitem slicers, validate that we are in fact integers
+            if is_index_slice:
+                # In this case the _validate_indexer checks below are redundant
+                return key
+            elif self.dtype.kind in "iu":
+                # Note: these checks are redundant if we know is_index_slice
+                self._validate_indexer("slice", key.start, "getitem")
+                self._validate_indexer("slice", key.stop, "getitem")
+                self._validate_indexer("slice", key.step, "getitem")
+                return key
+
+        # convert the slice to an indexer here
 
         # special case for interval_dtype bc we do not do partial-indexing
         #  on integer Intervals when slicing
@@ -4098,17 +4110,6 @@ class Index(IndexOpsMixin, PandasObject):
             self.dtype, IntervalDtype
         )
         is_positional = is_index_slice and ints_are_positional
-
-        if kind == "getitem":
-            # called from the getitem slicers, validate that we are in fact integers
-            if is_index_slice or is_integer_dtype(self.dtype):
-                # Note: these checks are redundant if we know is_index_slice
-                self._validate_indexer("slice", key.start, "getitem")
-                self._validate_indexer("slice", key.stop, "getitem")
-                self._validate_indexer("slice", key.step, "getitem")
-                return key
-
-        # convert the slice to an indexer here
 
         # if we are mixed and have integers
         if is_positional:
@@ -4141,7 +4142,7 @@ class Index(IndexOpsMixin, PandasObject):
     @final
     def _raise_invalid_indexer(
         self,
-        form: str_t,
+        form: Literal["slice", "positional"],
         key,
         reraise: lib.NoDefault | None | Exception = lib.no_default,
     ) -> None:
@@ -6369,14 +6370,17 @@ class Index(IndexOpsMixin, PandasObject):
         return ensure_index(target)
 
     @final
-    def _validate_indexer(self, form: str_t, key, kind: str_t) -> None:
+    def _validate_indexer(
+        self,
+        form: Literal["positional", "slice"],
+        key,
+        kind: Literal["getitem", "iloc"],
+    ) -> None:
         """
         If we are positional indexer, validate that we have appropriate
         typed bounds must be an integer.
         """
-        assert kind in ["getitem", "iloc"]
-
-        if key is not None and not is_integer(key):
+        if not lib.is_int_or_none(key):
             self._raise_invalid_indexer(form, key)
 
     def _maybe_cast_slice_bound(self, label, side: str_t):