… pandas/_libs/lib.pyi