@alippai 2023-03-31
Interesting, why is the df faster than the series?